Make galpy.df a proper subpackage, by moving galpy/df_src/ --> galpy/df/ and galpy/df.py --> galpy/df/__init.py; various other necessary adjustments
