+++ MANIFEST.in
@@ -1,6 +1,6 @@
 include README.rst README.dev README.nemo LICENSE HISTORY.txt AUTHORS.txt
 include gsl-config.bat
-include galpy/df_src/data/*.sav
+include galpy/df/data/*.sav
 include galpy/actionAngle_src/actionAngle_c_ext/*.h
 include galpy/actionAngle_src/actionAngleTorus_c_ext/*.h
 include galpy/orbit_src/orbit_c_ext/*.h
