+++ README.rst
@@ -143,7 +143,7 @@ profiles more closely (see `1999AJ....118.1201D
 these corrections is expensive, and a large set of precalculated
 corrections can be found `here
 <http://github.com/downloads/jobovy/galpy/galpy-dfcorrections.tar.gz>`__
-\[tar.gz archive\]. Install these by downloading them and unpacking them into the galpy/df_src/data directory before running the setup.py installation. E.g.::
+\[tar.gz archive\]. Install these by downloading them and unpacking them into the galpy/df/data directory before running the setup.py installation. E.g.::
 
    curl -O https://github.s3.amazonaws.com/downloads/jobovy/galpy/galpy-dfcorrections.tar.gz
-   tar xvzf galpy-dfcorrections.tar.gz -C ./galpy/df_src/data/
+   tar xvzf galpy-dfcorrections.tar.gz -C ./galpy/df/data/
