+++ galpy/df/jeans.py
