+++ setup.py
@@ -298,9 +298,9 @@ setup(name='galpy',
       url='http://github.com/jobovy/galpy',
       package_dir = {'galpy/': ''},
       packages=['galpy','galpy/orbit_src','galpy/potential_src',
-                'galpy/df_src','galpy/util','galpy/snapshot_src',
+                'galpy/df','galpy/util','galpy/snapshot_src',
                 'galpy/actionAngle_src'],
-      package_data={'galpy/df_src':['data/*.sav'],
+      package_data={'galpy/df':['data/*.sav'],
                     "": ["README.rst","README.dev","LICENSE","AUTHORS.rst"]},
       include_package_data=True,
       install_requires=['numpy>=1.7','scipy','matplotlib','pytest','six'],
